requirements.txt
app==0.0.1
attrs==23.1.0
blinker==1.6.2
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
coverage==7.3.2
crypto==1.4.1
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Dance==6.2.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-OAuthlib==0.9.6
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
MyApplication==0.1.0
Naked==0.1.32
oauthlib==2.1.0
outcome==1.3.0.post0
Pillow==10.0.1
psycopg==3.0.12
psycopg2-binary==2.9.3
PyJWT==2.8.0
PySocks==1.7.1
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.1.0
selenium==4.14.0
shellescape==3.8.1
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.20
testcase==0.1.0
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
urllib3==2.0.4
URLObject==2.4.3
utils==1.0.1
Werkzeug==3.0.0
wsproto==1.2.0
